UC9 commit 3 - Add user function
Added the ability for the user to add an account. They have to fill in their name, email and password. The user gets added to the system and is now able to log in with their newly made account.

The system shows feedback on their input, is the email correct, are all entry fields filled.

Password gets hashed with the HashPassword function.